Add allow-exit config
With the addition of
"allow-exit": [ "redcap_custom_verify_username" ],
to the config.json, you can then call exit() inside of your hook without triggering a EM error.  I know this has downsides and could prevent a second EM using the same hook from ever getting a chance to fire, but I feel like it is a necessary evil if we are going to be able to use some of these hooks to their fullest (such as the one mentioned above which no longer supports return values with EMs.